requirements.txt
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<2.4.0
scipy>=1.10.0
numexpr>=2.8.7
bottleneck>=1.3.7

librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
tqdm>=4.66.0
yt-dlp>=2024.3.10
requests>=2.31.0

# Web UI
gradio>=4.0.0
fastapi>=0.115.0
pydantic>=2.0.0,<2.12.0

# Separation / DSP (Demucs requires torch; install matching wheel for your platform)
demucs>=4.0.0
torch>=2.1.0

# ML features
scikit-learn>=1.3.0

# Lyrics extraction (optional; Whisper + fallbacks)
openai-whisper>=20231117
SpeechRecognition>=3.10.0
pocketsphinx>=5.0.0

